Refactored code for creating the apps signature custom marker
